requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
Flask==2.0.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
jsonapi-requests==0.6.2
MarkupSafe==2.0.1
paho-mqtt==1.5.1
parso==0.8.2
requests==2.25.1
six==1.16.0
tenacity==7.0.0
urllib3==1.26.5
Werkzeug==2.0.1
